+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
-- Update DESIGN.md
+- Update docs.
+- Use [cybozu-go/netutil](https://github.com/cybozu-go/netutil).
 
 [Unreleased]: https://github.com/cybozu-go/aptutil/compare/v1.0.0...HEAD